Move common test utilities
